chore: Moved artefact inference script to doctr/scripts
